requirements.txt
yfinance>=0.2.28
fredapi>=0.5.1  # Optional: for FRED data (free API key required)
pandas>=2.0.0
pyarrow>=10.0.0  # Required for Parquet cache support
numpy>=1.24.0
hmmlearn>=0.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.66.0
pytest>=7.4.0  # Testing framework
pytest-mock>=3.11.0  # Mocking utilities for pytest